refactor: 스토리 파일들이 컴포넌트와 같은 곳에 위치하도록 폴더 구조 변경
